+++ packages/provider/src/api/blacklistRequestInspector.ts
@@ -34,7 +34,9 @@ export class BlacklistRequestInspector {
 	): Promise<void> {
 		const rawIp = request.ip || "";
 
-		request.logger.debug("JA4", request.ja4);
+		request.logger.debug(() => ({
+			data: { ja4: request.ja4 },
+		}));
 
 		const shouldAbortRequest = await this.shouldAbortRequest(
 			request.url,
@@ -68,11 +70,14 @@ export class BlacklistRequestInspector {
 
 		// block if no IP is present
 		if (!rawIp) {
-			logger.info("Request without IP", {
-				requestedRoute: requestedRoute,
-				requestHeaders: requestHeaders,
-				requestBody: requestBody,
-			});
+			logger.info(() => ({
+				data: {
+					requestedRoute: requestedRoute,
+					requestHeaders: requestHeaders,
+					requestBody: requestBody,
+				},
+				msg: "Request without IP",
+			}));
 
 			return true;
 		}
@@ -102,7 +107,10 @@ export class BlacklistRequestInspector {
 
 			return AccessPolicyType.Block === accessPolicy?.type;
 		} catch (err) {
-			logger.error("Block Middleware Error:", err);
+			logger.error(() => ({
+				err,
+				msg: "Block Middleware Error",
+			}));
 
 			return true;
 		}