+++ packages/provider/src/api/block.ts
@@ -12,13 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { getLogger } from "@prosopo/common";
+import { getLogger, parseLogLevel } from "@prosopo/common";
 import type { ProviderEnvironment } from "@prosopo/types-env";
 import { createAccessPolicyResolver } from "@prosopo/user-access-policy";
 import { BlacklistRequestInspector } from "./blacklistRequestInspector.js";
 
 export const blockMiddleware = (providerEnvironment: ProviderEnvironment) => {
-	const logLevel = providerEnvironment.config.logLevel;
+	const logLevel = parseLogLevel(providerEnvironment.config.logLevel);
 	const logger = getLogger(logLevel, "blockMiddleware");
 
 	const userAccessRulesStorage = providerEnvironment