+++ packages/provider/src/api/verify.ts
@@ -105,7 +105,7 @@ export function prosopoVerifyRouter(env: ProviderEnvironment): Router {
 						ip,
 					);
 
-				req.logger.debug(response);
+				req.logger.debug(() => ({ data: { response } }));
 				const verificationResponse: ImageVerificationResponse =
 					tasks.imgCaptchaManager.getVerificationResponse(
 						response[ApiParams.verified],
@@ -116,7 +116,7 @@ export function prosopoVerifyRouter(env: ProviderEnvironment): Router {
 					);
 				res.json(verificationResponse);
 			} catch (err) {
-				req.logger.error({ err, body: req.body });
+				req.logger.error(() => ({ err, data: { body: req.body } }));
 				return next(
 					new ProsopoApiError("API.BAD_REQUEST", {
 						context: { code: 500, siteKey: req.body.dapp, user: req.body.user },
@@ -209,7 +209,7 @@ export function prosopoVerifyRouter(env: ProviderEnvironment): Router {
 				return res.json(verificationResponse);
 			} catch (err) {
 				console.error("\nError in verifyPowCaptchaSolution:", err);
-				req.logger.error({ err, body: req.body });
+				req.logger.error(() => ({ err, data: { body: req.body } }));
 				return next(
 					new ProsopoApiError("API.BAD_REQUEST", {
 						context: { code: 500, error: err },