+++ packages/user-access-policy/src/tests/testLogger.ts
@@ -12,6 +12,27 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { getLogger } from "@prosopo/common";
+import type { Logger } from "@prosopo/common";
+import { vi } from "vitest";
 
-export const testLogger = getLogger("debug", "test");
+const loggerMockedInstance: Logger = {
+	trace: vi.fn(),
+	debug: vi.fn(),
+	info: vi.fn(),
+	warn: vi.fn(),
+	error: vi.fn(),
+	fatal: vi.fn(),
+	log: vi.fn(),
+	setLogLevel: vi.fn(),
+	getLogLevel: vi.fn(),
+	with: vi.fn().mockReturnThis(),
+	getScope: vi.fn().mockReturnValue("test-scope"),
+	getPretty: vi.fn().mockReturnValue(false),
+	setPretty: vi.fn(),
+	getPrintStack: vi.fn().mockReturnValue(false),
+	setPrintStack: vi.fn(),
+	getFormat: vi.fn().mockReturnValue("json"),
+	setFormat: vi.fn(),
+};
+
+export { loggerMockedInstance };