+++ dev/vite-plugin-watch-workspace/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @prosopo/vite-plugin-watch-workspace
 
+## 2.6.1
+### Patch Changes
+
+- Updated dependencies [86c22b8]
+  - @prosopo/config@2.6.1
+
 ## 2.6.0
 
 ### Minor Changes