+++ packages/account/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @prosopo/account
 
+## 2.7.2
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/util-crypto@13.5.1
+  - @prosopo/keyring@2.8.1
+  - @prosopo/common@2.7.2
+  - @prosopo/types@2.9.1
+  - @prosopo/util@2.6.1
+
 ## 2.7.1
 ### Patch Changes