+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @prosopo/cli
 
+## 2.7.3
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/api-express-router@2.6.4
+  - @prosopo/user-access-policy@2.6.7
+  - @prosopo/util-crypto@13.5.1
+  - @prosopo/provider@2.12.1
+  - @prosopo/keyring@2.8.1
+  - @prosopo/common@2.7.2
+  - @prosopo/dotenv@2.6.4
+  - @prosopo/types@2.9.1
+  - @prosopo/api@2.7.1
+  - @prosopo/env@2.7.2
+  - @prosopo/config@2.6.1
+  - @prosopo/util@2.6.1
+
 ## 2.7.2
 ### Patch Changes