+++ packages/common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prosopo/common
 
+## 2.7.2
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/util-crypto@13.5.1
+
 ## 2.7.1
 ### Patch Changes