+++ packages/datasets-fs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @prosopo/datasets-fs
 
+## 2.7.2
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/common@2.7.2
+  - @prosopo/types@2.9.1
+  - @prosopo/util@2.6.1
+
 ## 2.7.1
 ### Patch Changes