+++ packages/provider/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @prosopo/provider
 
+## 2.12.1
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/api-express-router@2.6.4
+  - @prosopo/user-access-policy@2.6.7
+  - @prosopo/types-database@2.7.5
+  - @prosopo/util-crypto@13.5.1
+  - @prosopo/api-route@2.6.4
+  - @prosopo/types-env@2.7.2
+  - @prosopo/database@2.6.8
+  - @prosopo/datasets@2.7.2
+  - @prosopo/keyring@2.8.1
+  - @prosopo/common@2.7.2
+  - @prosopo/types@2.9.1
+  - @prosopo/env@2.7.2
+  - @prosopo/config@2.6.1
+  - @prosopo/util@2.6.1
+
 ## 2.12.0
 ### Minor Changes